refactor(social): move SocialError to dedicated error.rs
